refactor: viewController에 존재하는 decodeData()메서드객체화를 분리
